deleted useless jsx files, and moved some jsx files in pages directory
