--- client/src/pages/Actor.jsx
@@ -1,7 +1,0 @@
-import ActorList from "../components/ActorList";
-
-function Actor() {
-  return <ActorList />;
-}
-
-export default Actor;
