--- client/src/pages/Movie.jsx
@@ -1,7 +1,0 @@
-import MovieDetails from "../components/MovieDetails";
-
-function Movie() {
-  return <MovieDetails />;
-}
-
-export default Movie;
