--- client/src/pages/MyActor.jsx
@@ -1,7 +1,0 @@
-import ActorDetails from "../components/ActorDetails";
-
-function MyActor() {
-  return <ActorDetails />;
-}
-
-export default MyActor;
